Semihosting requires cooperation with the execution environment
Semihosting doesn't work in a vacuum, it needs something to implement
this ABI from outside the execution environment in question.  This
specifically came up as a "why don't ebreaks from U-mode result in
semihosting traps in QEMU?" discussion, I couldn't find anything to
point to saying that's the expected behavior.

I'm not sure this would really help anyone who's confused about this,
but at least it's something to point to next time.