Relax an overly-strict requirement for systems with paging
Just because a system has paging does not mean it is currently in use.
Running a simple RTOS on a full Sv39-supporting core that doesn't enable
it is no different to running on one without Sv39 support, and crossing
a page boundary is totally fine. Similarly, firmware running in M-mode
on a Unix-capable core with a full OS running in S-mode with paging can
still not care about paging, since the firmware's own program counter is
not subject to translation.